fix(package): exposed missing implicitBaseUrlSymbol, readTsConfig, writeTsConfig, writeTsConfigSync functions
