+++ .npmrc
@@ -6,6 +6,8 @@ strict-peer-dependencies=false
 public-hoist-pattern[]=secure-password
 public-hoist-pattern[]=zod
 
+public-hoist-pattern[]=@visulima/packem
+
 # Needed for crud package to work properly. Don't remove the lines.
 
 public-hoist-pattern[]=*prisma*
