+++ packages/package/__tests__/unit/read-tsconfig-node-modules.test.ts
@@ -14,7 +14,7 @@ import stripAnsi from "strip-ansi";
 import { temporaryDirectory } from "tempy";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { readTsConfig } from "../../src/read-tsconfig";
+import { readTsConfig } from "../../src/tsconfig/read-tsconfig";
 import { esc, getTscTsconfig } from "../helpers";
 
 describe("node_modules", () => {
