+++ packages/package/packem.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from "@visulima/packem/config";
+import transformer from "@visulima/packem/transformer/esbuild";
+
+// eslint-disable-next-line import/no-unused-modules
+export default defineConfig({
+    rollup: {
+        license: {
+            path: "./LICENSE.md",
+        },
+        shim: true,
+    },
+    transformer,
+});
