+++ packages/package/tsconfig.eslint.json
@@ -4,5 +4,5 @@
     "compilerOptions": {
         "baseUrl": "."
     },
-    "include": ["__docs__/**/*", "__tests__/**/*", "src/**/*", "*.d.ts", "tsup.config.ts"]
+    "include": ["__docs__/**/*", "__tests__/**/*", "src/**/*", "*.d.ts", "packem.config.ts"]
 }
