--- packages/package/tsup.config.ts
@@ -1,5 +1,0 @@
-import { createConfig } from "../../tools/get-tsup-config";
-
-const config = createConfig({});
-
-export default config;
